Fix getAxis implementation in Revolute and Prismatic Joint
The direction of the axis in a joint depends on which body
you consider parent and which body you consider child .
See "Minimal formulation of joint motion for biomechanisms",
Section 2.4 and Traversaro's PhD Thesis, Chapter 3.

This is a brief change, but the method documentation should
be appropriatly expanded to document this behaviour.